+++ src/main/java/uz/mediasolutions/mdeliveryservice/payload/ProductWebDTO.java
@@ -17,7 +17,7 @@ public class ProductWebDTO {
 
     private String imageUrl;
 
-    private Double price;
+    private float price;
 
     private boolean oneVariation;
