+++ src/main/java/uz/mediasolutions/mdeliveryservice/payload/VariationResDTO.java
@@ -17,9 +17,9 @@ public class VariationResDTO {
 
     private MeasureUnitDTO measureUnit;
 
-    private Double measure;
+    private float measure;
 
-    private Double price;
+    private float price;
 
     private Integer count;
